WiP: Plugin architecture study.
