+++ scripts/store/configureStore.js
@@ -2,7 +2,7 @@ import { createStore, applyMiddleware, compose } from "redux";
 import { hashHistory } from "react-router";
 import { routerMiddleware } from "react-router-redux";
 import createSagaMiddleware from "redux-saga";
-import rootReducer from "../reducers";
+import createRootReducer from "../reducers";
 import rootSaga from "../sagas";
 
 
@@ -13,13 +13,12 @@ const finalCreateStore = compose(
   window.devToolsExtension ? window.devToolsExtension() : f => f
 )(createStore);
 
-export default function configureStore(initialState) {
-  const store = finalCreateStore(rootReducer, initialState);
+export default function configureStore(initialState, plugins=[]) {
+  const store = finalCreateStore(createRootReducer(plugins), initialState);
   // Every saga will receive the store getState() function as first argument
   // by default.
   // This allows sagas to share the same signature and access the state
   // consistently.
-  const getState = store.getState.bind(store);
-  sagaMiddleware.run(rootSaga.bind(null, getState));
+  sagaMiddleware.run(rootSaga, store.getState.bind(store), plugins);
   return store;
 }
